Refactor backend (#68)
* "ajout possibilité de récuperer les circuits à partir d'une ville"

* commit

* Set up security verification related to the user decoded token in the context and refacto circuit and poi logic

---------

+++ src/services/poi.service.ts
@@ -2,19 +2,22 @@ import "reflect-metadata";
 import { QueryFailedError } from "typeorm";
 
 import Poi from "../entities/PointOfInterest.entity";
+import Tag from "../entities/Tag.entity";
 import { CustomError } from "../utils/errors/CustomError.utils.error";
 import { InternalServerError } from "../utils/errors/interfaces.utils.error";
 import { formatString } from "../utils/string.utils";
 import { 
   PoiErrorsFlag, 
   handlePoiError, 
-  handlePoiObjectError, 
-  handlePoiTagObjectError 
+  handlePoiObjectError
 } from "../utils/errors/handleError/poi.utils.error";
 import { PoiValidator } from "../validators/entities/poi.validator.entity";
-import { TagValidator } from "../validators/entities/tag.validator.entity";
 import { PoiRepository } from "../repositories/poi.repository";
 import { TagRepository } from "../repositories/tag.repository";
+import { CityRepository } from "../repositories/city.repository";
+import { TypeRepository } from "../repositories/type.repository";
+import { MyAppContext } from "../types/context";
+import User from "../entities/User.entity";
 
 
 /**
@@ -24,15 +27,13 @@ import { TagRepository } from "../repositories/tag.repository";
  */
 export const getAll = async (): Promise<Array<Poi>> => {
   try {
-    const allPoi: Poi[] = await PoiRepository.find({
-      relations: {
-          tags: true,
-      }
+    const allPoi: Poi[] | null = await PoiRepository.find({
+      relations: ["city"]
     });
-    
-    return allPoi;
+    return allPoi ? allPoi : [];
   } catch (e) {
     throw new CustomError(
+      
       new InternalServerError(), 
       `Problème de connexion interne, les points d'intêret n'ont pas été chargés`
     );
@@ -47,8 +48,12 @@ export const getAll = async (): Promise<Array<Poi>> => {
  */
 export const getById = async (id: number): Promise<Poi> => {
   try {
-    const isPoiExist = await PoiRepository.findOneBy({id});
-    if (isPoiExist) return isPoiExist;
+    const poiInDb = await PoiRepository.findOne({
+      where: { id: id },
+      relations: ['city'],
+    });
+    
+    if (poiInDb) return poiInDb;
     else throw new Error(PoiErrorsFlag.ID_NOT_FOUND);
   } catch (e) {
     if (e instanceof Error) handlePoiError(e, null);    
@@ -83,35 +88,53 @@ export const getByName = async (name: string): Promise<Poi> => {
 /**
  * Create and return a point of interest
  * @param {PoiValidator} data point of interest object to create 
+ * @param {MyAppContext} ctx the app context with the decoded token
  * @returns point of interest the created poi 
  * @throws Error: 500 Internal Server Error | 422 Unprocessable Entity
 */
 export const create = async (
-  data: PoiValidator
+  data: PoiValidator,
+  ctx: MyAppContext
 ): Promise<Poi> => {
   data.name = formatString(data.name);
-  let tagIsNotInDB = null;
+  let tagsInDb: Tag[] | null = [];
   
   try {
+    const cityInDb = await CityRepository.findOneBy({ id: data.cityId });
+    // Check if city is present in database
+    if (!cityInDb) throw new Error(PoiErrorsFlag.CITY_NOT_IN_DB);
+
+    // Check if the request sender is the user attached to the city
+    if (!verifyIfTwoUsersAreIdentical(ctx, cityInDb.user)) throw new Error(PoiErrorsFlag.USER_NOT_AUTHORIZED_CREATE);
+
+    const typeInDb = await TypeRepository.findOneBy({ id: data.typeId }); 
+    // Check if type is present in database
+    if (!typeInDb) throw new Error(PoiErrorsFlag.TYPE_NOT_IN_DB);
+
     // Check if the localisation already exist
     const isLocalisationAlreadyExist = await PoiRepository.findByLatitudeAndByLongitude(data.latitude, data.longitude);        
     if (isLocalisationAlreadyExist) throw new Error(PoiErrorsFlag.LOCALISATION_ALREADY_USED);
     
+
     // Check if all tags are present in database
     if (data.tags !== null && data.tags.length > 0) {
-      tagIsNotInDB = await checkIfTagsAllExist(data.tags);
+      tagsInDb = await checkIfTagsAllExist(data.tags);
     } 
-    if (tagIsNotInDB) throw new Error(PoiErrorsFlag.TAG_NOT_IN_DB);
+    if (!tagsInDb) throw new Error(PoiErrorsFlag.TAG_NOT_IN_DB);
     
-    const createdPoi = await PoiRepository.save(data);    
-    return createdPoi;
-  } catch (e) {    
-    if (
-      e instanceof Error && 
-      e.message === PoiErrorsFlag.TAG_NOT_IN_DB &&
-      tagIsNotInDB !== null
-    ) handlePoiTagObjectError(e, tagIsNotInDB);
+    const newPoi = new Poi();
+    newPoi.name = data.name;
+    newPoi.address = data.address;
+    newPoi.latitude = data.latitude;
+    newPoi.longitude = data.longitude;
+    newPoi.picture = data.picture;
+    newPoi.type = typeInDb;
+    newPoi.city = cityInDb;
+    newPoi.tags = tagsInDb;
 
+    return await PoiRepository.save(newPoi);    
+  } catch (e) {    
+    if (e instanceof Error ) handlePoiError(e, data);
     if (e instanceof QueryFailedError || e instanceof Error) {
       handlePoiObjectError(e, data);
     } throw new CustomError(
@@ -124,45 +147,70 @@ export const create = async (
 /**
  * Update and return a point of interest
  * @param {PoiValidator} data poi object to update
+ * @param {MyAppContext} ctx the app context with the decoded token
  * @returns updated poi
  * @throws Error: 500 Internal Server Error | 404 Not Found | 422 Unprocessable Entity
  */
 export const update = async (
-  data: PoiValidator
+  data: PoiValidator,
+  ctx: MyAppContext
 ): Promise<Poi> => {
   data.name = formatString(data.name);
-  let tagIsNotInDB = null;
+  let tagsInDb: Tag[] | null = [];
   
   try {
     // Check if the id is present in database
-    const isIdExistInDB = await PoiRepository.findOneBy({ id: data.id });
-    if (!isIdExistInDB) throw new Error(PoiErrorsFlag.ID_NOT_FOUND); 
+    const poiInDB = await PoiRepository.findOne({
+      where: { id: data.id },
+      relations: ['city'],
+    });
+    if (!poiInDB) throw new Error(PoiErrorsFlag.ID_NOT_FOUND); 
+
+    // Check if city is present in database
+    const cityInDb = await CityRepository.findOneBy({ id: data.cityId });
+    if (!cityInDb) throw new Error(PoiErrorsFlag.CITY_NOT_IN_DB);
+
+    // Check if request sender is the user attached to the city
+    if (
+      !verifyIfTwoUsersAreIdentical(ctx, cityInDb.user) ||
+      !verifyIfTwoUsersAreIdentical(ctx, poiInDB.city.user)
+    ) throw new Error(PoiErrorsFlag.USER_NOT_AUTHORIZED_UPDATE);
+
+    // Check if type is present in database
+    const typeInDb = await TypeRepository.findOneBy({ id: data.typeId }); 
+    if (!typeInDb) throw new Error(PoiErrorsFlag.TYPE_NOT_IN_DB);
     
     // Check if the latitude and longitude are in database for all different id
-    if (isIdExistInDB.latitude !== data.latitude || isIdExistInDB.longitude !== data.longitude) {
-      const isLocationAlreadyExist = await PoiRepository.findByLatitudeAndByLongitudeIfNotID(
-        data.id, 
-        data.latitude, 
-        data.longitude
-      );
+    if (poiInDB.latitude !== data.latitude || poiInDB.longitude !== data.longitude) {
+      const isLocationAlreadyExist = await PoiRepository
+        .findByLatitudeAndByLongitudeIfNotID(
+          data.id, 
+          data.latitude, 
+          data.longitude
+        );
       if (isLocationAlreadyExist) throw new Error(PoiErrorsFlag.LOCALISATION_ALREADY_USED);
     }
     
     // Check if all tags are present in database
     if (data.tags !== null && data.tags.length > 0) {
-      tagIsNotInDB = await checkIfTagsAllExist(data.tags);
+      tagsInDb = await checkIfTagsAllExist(data.tags);
     } 
-    if (tagIsNotInDB) throw new Error(PoiErrorsFlag.TAG_NOT_IN_DB);
+    if (!tagsInDb) throw new Error(PoiErrorsFlag.TAG_NOT_IN_DB);
+    
+    const newPoi = new Poi();
+    newPoi.name = data.name;
+    newPoi.address = data.address;
+    newPoi.latitude = data.latitude;
+    newPoi.longitude = data.longitude;
+    newPoi.picture = data.picture;
+    newPoi.type = typeInDb;
+    newPoi.city = cityInDb;
+    // Update poi tags
+    (data.tags?.length > 0) ? newPoi.tags = [...tagsInDb] : newPoi.tags = [];
 
-    (data.tags?.length > 0) ? isIdExistInDB.tags = [...data.tags] : isIdExistInDB.tags = [];
-
-    return await PoiRepository.save({ ...isIdExistInDB, ...data });
+    return await PoiRepository.save({ ...poiInDB, ...newPoi });
   } catch (e) {
-    if (
-      e instanceof Error && 
-      e.message === PoiErrorsFlag.TAG_NOT_IN_DB &&
-      tagIsNotInDB !== null
-    ) handlePoiTagObjectError(e, tagIsNotInDB);
+    if (e instanceof Error) handlePoiError(e, data);
 
     if (e instanceof QueryFailedError || e instanceof Error) {
       handlePoiObjectError(e, data);
@@ -179,12 +227,19 @@ export const update = async (
  * @returns deleted poi 
  * @throws Error: 500 Internal Server Error | 404 Not Found
  */
-export const deletePoi = async (id: number): Promise<Poi> => {
+export const deletePoi = async (id: number, ctx: MyAppContext): Promise<Poi> => {
   try {
-    const poiToRemove = await PoiRepository.findOneBy({id});
+    const poiToRemove = await PoiRepository.findOne({
+      where: { id: id },
+      relations: ['city'],
+    });
+
+    // Check if the request sender is the user attached to the city
+    // If yes we can remove the poi
     if (poiToRemove) {
-      await PoiRepository.remove(poiToRemove);
-      return poiToRemove;
+      if (!verifyIfTwoUsersAreIdentical(ctx, poiToRemove?.city?.user)) throw new Error(PoiErrorsFlag.USER_NOT_AUTHORIZED_DELETE);
+      
+      return await PoiRepository.remove(poiToRemove);
     } else throw new Error(PoiErrorsFlag.ID_NOT_FOUND);
   } catch (e) {
     if (e instanceof Error) handlePoiError(e, null);    
@@ -195,11 +250,35 @@ export const deletePoi = async (id: number): Promise<Poi> => {
   }
 };
 
-const checkIfTagsAllExist = async (tags: TagValidator[]): Promise<null | string> => {
-  for (let i = 0; i < tags.length; i++) {
-    const id = tags[i].id;
-    const isTagInDB = await TagRepository.findOneBy({id});
-    if (!isTagInDB) return tags[i].name;
+/**
+ * Return a tags array from database or null if a tag doesn't exist
+ * @param {number[]} tagIds the tag ids array
+ * @returns Tag[] | null 
+ * @throws Error: 500 Internal Server Error | 404 Not Found
+*/
+const checkIfTagsAllExist = async (tagIds: number[]): Promise<Tag[] | null> => {
+  let tagsInDb: Tag[] = [];
+
+  for (let i = 0; i < tagIds.length; i++) {
+    const id = tagIds[i];
+    const tagInDB = await TagRepository.findOneBy({id});
+
+    if (!tagInDB) return null;
+    else tagsInDb.push(tagInDB);
   }
-  return null;
+  return tagsInDb;
+}
+
+/**
+ * Check if the user in the context is identical to the user required for the action
+ * @param {MyAppContext} senderCtx the actual context for the request sender
+ * @param {User} userToCompare the user to compare with
+ * @returns true if the two users are identical | false otherwise 
+*/
+const verifyIfTwoUsersAreIdentical = (senderCtx: MyAppContext, userToCompare: User): boolean => {
+  return (
+    senderCtx.user.id === userToCompare.id &&
+    senderCtx.user.username === userToCompare.username &&
+    senderCtx.user.email === userToCompare.email
+  );
 }
